requirements.txt
uvicorn[standard]==0.42.0
python-dotenv==1.2.2
requests==2.32.5
yt-dlp==2026.3.17
opencv-python==4.13.0.92
keras-facenet==0.3.2
tensorflow==2.21.0
numpy==2.4.3
librosa==0.11.0
soundfile==0.13.1
torch==2.10.0
torchaudio==2.10.0
transformers==5.3.0